refactor(backend): extract followup pool collectors + digest (813 -> 487 lines)
