+++ backend/app/services/followup_engine.py
@@ -198,320 +198,6 @@ def _get_interval(priority_settings: dict | None, level: str) -> int:
     return settings.get(level, DEFAULT_PRIORITY_SETTINGS.get(level, TIME_BASED_INACTIVITY_DAYS))
 
 
-# -----------------------------------------------------------------------
-# Pool A — Active relationships
-# -----------------------------------------------------------------------
-
-
-async def _collect_pool_a_candidates(
-    user_id: uuid.UUID,
-    db: AsyncSession,
-    now: datetime,
-    queued_contact_ids: set[uuid.UUID],
-    priority_settings: dict | None = None,
-    dormancy_days: int = DORMANCY_THRESHOLD_DAYS,
-) -> dict[uuid.UUID, _Candidate]:
-    """Collect candidates from triggers 1-4 for active (non-dormant) contacts.
-    ``dormancy_days`` is the user's configured upper bound — contacts whose
-    last interaction is older than this fall through to Pool B."""
-    candidates: dict[uuid.UUID, _Candidate] = {}
-    dormancy_cutoff = now - timedelta(days=dormancy_days)
-
-    # ------------------------------------------------------------------
-    # Trigger 1: Time-based — no interaction in N+ days, score < 4, not dormant
-    # ------------------------------------------------------------------
-    for level in ("high", "medium", "low"):
-        interval_days = _get_interval(priority_settings, level)
-        cutoff_time = now - timedelta(days=interval_days)
-        time_based_result = await db.execute(
-            select(Contact).where(
-                Contact.user_id == user_id,
-                _not_2nd_tier,
-                _has_channel,
-                _has_interactions,
-                _not_archived,
-                Contact.priority_level == level,
-                Contact.relationship_score < TIME_BASED_MAX_SCORE,
-                Contact.last_interaction_at < cutoff_time,
-                Contact.last_interaction_at >= dormancy_cutoff,
-                Contact.interaction_count >= MIN_INTERACTIONS_FOR_SUGGESTION.get(level, 3),
-            )
-        )
-        for contact in time_based_result.scalars().all():
-            if contact.id in queued_contact_ids:
-                continue
-            days_since = _days_since(contact.last_interaction_at, now)
-            if contact.relationship_score == 0 and days_since > STALE_CONTACT_DAYS:
-                continue
-            priority_score = compute_priority(contact.interaction_count, days_since, False)
-            if contact.id not in candidates or priority_score > candidates[contact.id].priority:
-                candidates[contact.id] = _Candidate(
-                    contact=contact, trigger_type="time_based", priority=priority_score, pool="A",
-                )
-
-    # ------------------------------------------------------------------
-    # Trigger 2: Event-based — DetectedEvents in last 7 days, confidence > 0.7, not dormant
-    # Skip contacts contacted in the last COOLING_RECENT_DAYS — even with
-    # a noteworthy event, a follow-up 5 days after the last conversation
-    # is too soon.
-    # ------------------------------------------------------------------
-    event_cutoff = now - timedelta(days=EVENT_BASED_WINDOW_DAYS)
-    cooling_cutoff = now - timedelta(days=COOLING_RECENT_DAYS)
-    events_result = await db.execute(
-        select(DetectedEvent, Contact)
-        .join(Contact, DetectedEvent.contact_id == Contact.id)
-        .where(
-            Contact.user_id == user_id,
-            _not_2nd_tier,
-            _has_channel,
-            _has_interactions,
-            _not_archived,
-            Contact.last_interaction_at >= dormancy_cutoff,
-            Contact.last_interaction_at < cooling_cutoff,  # respect cooldown
-            DetectedEvent.detected_at >= event_cutoff,
-            DetectedEvent.confidence > EVENT_CONFIDENCE_THRESHOLD,
-        )
-        .order_by(DetectedEvent.confidence.desc())
-    )
-    for event, contact in events_result.all():
-        if contact.id in queued_contact_ids:
-            continue
-        days_since = _days_since(contact.last_interaction_at, now)
-        priority = compute_priority(contact.interaction_count, days_since, True)
-        if contact.id not in candidates or priority > candidates[contact.id].priority:
-            candidates[contact.id] = _Candidate(
-                contact=contact, trigger_type="event_based", event=event, priority=priority, pool="A",
-            )
-
-    # ------------------------------------------------------------------
-    # Trigger 3: Scheduled — last_followup_at > N days ago, not dormant
-    # ------------------------------------------------------------------
-    for level in ("high", "medium", "low"):
-        interval_days = _get_interval(priority_settings, level)
-        scheduled_cutoff = now - timedelta(days=interval_days)
-        scheduled_result = await db.execute(
-            select(Contact).where(
-                Contact.user_id == user_id,
-                _not_2nd_tier,
-                _has_channel,
-                _has_interactions,
-                _not_archived,
-                Contact.priority_level == level,
-                Contact.last_followup_at.isnot(None),
-                Contact.last_followup_at < scheduled_cutoff,
-                Contact.last_interaction_at >= dormancy_cutoff,
-                Contact.interaction_count >= MIN_INTERACTIONS_FOR_SUGGESTION.get(level, 3),
-            )
-        )
-        for contact in scheduled_result.scalars().all():
-            if contact.id in queued_contact_ids:
-                continue
-            days_since = _days_since(contact.last_interaction_at, now)
-            if contact.relationship_score == 0 and days_since > STALE_CONTACT_DAYS:
-                continue
-            priority_score = compute_priority(contact.interaction_count, days_since, False)
-            if contact.id not in candidates or priority_score > candidates[contact.id].priority:
-                candidates[contact.id] = _Candidate(
-                    contact=contact, trigger_type="scheduled", priority=priority_score, pool="A",
-                )
-
-    # ------------------------------------------------------------------
-    # Trigger 4: Birthday — birthday within next 3 days (no dormancy filter)
-    # ------------------------------------------------------------------
-    today = now.date()
-    upcoming_mmdd = {(today + timedelta(days=d)).strftime("%m-%d") for d in range(4)}
-    birthday_result = await db.execute(
-        select(Contact).where(
-            Contact.user_id == user_id,
-            _not_2nd_tier,
-            _has_channel,
-            _not_archived,
-            Contact.birthday.isnot(None),
-        )
-    )
-    for contact in birthday_result.scalars().all():
-        if contact.id in queued_contact_ids:
-            continue
-        bday = contact.birthday.strip()
-        mmdd = bday[-5:]  # last 5 chars = "MM-DD"
-        if mmdd not in upcoming_mmdd:
-            continue
-        if contact.id not in candidates or 1500.0 > candidates[contact.id].priority:
-            candidates[contact.id] = _Candidate(
-                contact=contact, trigger_type="birthday", priority=1500.0, pool="A",
-            )
-
-    return candidates
-
-
-# -----------------------------------------------------------------------
-# Pool B — Dormant revival
-# -----------------------------------------------------------------------
-
-
-async def _collect_pool_b_candidates(
-    user_id: uuid.UUID,
-    db: AsyncSession,
-    now: datetime,
-    queued_contact_ids: set[uuid.UUID],
-    dormancy_days: int = DORMANCY_THRESHOLD_DAYS,
-) -> dict[uuid.UUID, _Candidate]:
-    """Collect candidates from dormant contacts worth reviving. ``dormancy_days``
-    is the lower bound: contacts dormant longer than this are Pool B candidates."""
-    candidates: dict[uuid.UUID, _Candidate] = {}
-    dormancy_cutoff = now - timedelta(days=dormancy_days)
-    hard_cap_cutoff = now - timedelta(days=HARD_CAP_DORMANCY_YEARS * 365)
-
-    # Build interaction span subquery (CTE)
-    span_subq = (
-        select(
-            Interaction.contact_id,
-            func.extract(
-                'epoch',
-                func.max(Interaction.occurred_at) - func.min(Interaction.occurred_at)
-            ).label('span_seconds'),
-        )
-        .group_by(Interaction.contact_id)
-        .subquery()
-    )
-
-    # Base dormant filter: last_interaction_at IS NULL or < dormancy cutoff
-    _is_dormant = or_(
-        Contact.last_interaction_at.is_(None),
-        Contact.last_interaction_at < dormancy_cutoff,
-    )
-
-    # Score-0 contacts are dead relationships — not worth reviving
-    _has_nonzero_score = Contact.relationship_score > 0
-
-    # ------------------------------------------------------------------
-    # Trigger B1 — Deep Dormant
-    # interaction_count >= 15 OR relationship_score >= 7
-    # last_interaction_at between 1-5 years ago
-    # ------------------------------------------------------------------
-    b1_min_cutoff = now - timedelta(days=365)  # at least 1 year dormant
-    b1_max_cutoff = now - timedelta(days=B1_MAX_DORMANCY_YEARS * 365)
-    b1_result = await db.execute(
-        select(Contact, span_subq.c.span_seconds)
-        .outerjoin(span_subq, Contact.id == span_subq.c.contact_id)
-        .where(
-            Contact.user_id == user_id,
-            _not_2nd_tier,
-            _has_channel,
-            _not_archived,
-            _has_nonzero_score,
-            Contact.last_interaction_at.isnot(None),
-            Contact.last_interaction_at < b1_min_cutoff,
-            Contact.last_interaction_at >= b1_max_cutoff,
-            or_(
-                Contact.interaction_count >= B1_MIN_INTERACTIONS,
-                Contact.relationship_score >= B1_MIN_SCORE,
-            ),
-        )
-    )
-    for contact, span_seconds in b1_result.all():
-        if contact.id in queued_contact_ids:
-            continue
-        span_days = float(span_seconds or 0) / 86400.0
-        priority = compute_priority_b(
-            contact.interaction_count, contact.relationship_score, span_days, False,
-        )
-        if contact.id not in candidates or priority > candidates[contact.id].priority:
-            candidates[contact.id] = _Candidate(
-                contact=contact, trigger_type="dormant_deep", priority=priority, pool="B",
-            )
-
-    # ------------------------------------------------------------------
-    # Trigger B2 — Mid-Dormant
-    # interaction_count >= 8 AND span >= 90 days
-    # last_interaction_at between 1-3 years ago
-    # ------------------------------------------------------------------
-    b2_max_cutoff = now - timedelta(days=B2_MAX_DORMANCY_YEARS * 365)
-    b2_result = await db.execute(
-        select(Contact, span_subq.c.span_seconds)
-        .outerjoin(span_subq, Contact.id == span_subq.c.contact_id)
-        .where(
-            Contact.user_id == user_id,
-            _not_2nd_tier,
-            _has_channel,
-            _not_archived,
-            _has_nonzero_score,
-            Contact.last_interaction_at.isnot(None),
-            Contact.last_interaction_at < b1_min_cutoff,
-            Contact.last_interaction_at >= b2_max_cutoff,
-            Contact.interaction_count >= B2_MIN_INTERACTIONS,
-            func.coalesce(span_subq.c.span_seconds, 0) >= B2_MIN_SPAN_DAYS * 86400,
-        )
-    )
-    for contact, span_seconds in b2_result.all():
-        if contact.id in queued_contact_ids:
-            continue
-        if contact.id in candidates:
-            continue  # B1 already captured this contact with higher priority
-        span_days = float(span_seconds or 0) / 86400.0
-        priority = compute_priority_b(
-            contact.interaction_count, contact.relationship_score, span_days, False,
-        )
-        if contact.id not in candidates or priority > candidates[contact.id].priority:
-            candidates[contact.id] = _Candidate(
-                contact=contact, trigger_type="dormant_mid", priority=priority, pool="B",
-            )
-
-    # ------------------------------------------------------------------
-    # Trigger B3 — Event Revival
-    # DetectedEvent in last 14 days, confidence > 0.7
-    # Must pass depth qualification (interaction_count >= 8 OR score >= 6 OR span >= 180d)
-    # Overrides 5-year hard cap (can surface ancient contacts)
-    # ------------------------------------------------------------------
-    b3_event_cutoff = now - timedelta(days=B3_EVENT_WINDOW_DAYS)
-    b3_result = await db.execute(
-        select(DetectedEvent, Contact, span_subq.c.span_seconds)
-        .join(Contact, DetectedEvent.contact_id == Contact.id)
-        .outerjoin(span_subq, Contact.id == span_subq.c.contact_id)
-        .where(
-            Contact.user_id == user_id,
-            _not_2nd_tier,
-            _has_channel,
-            _not_archived,
-            _has_nonzero_score,
-            _is_dormant,
-            DetectedEvent.detected_at >= b3_event_cutoff,
-            DetectedEvent.confidence > EVENT_CONFIDENCE_THRESHOLD,
-            or_(
-                Contact.interaction_count >= POOL_B_MIN_INTERACTIONS,
-                Contact.relationship_score >= POOL_B_MIN_SCORE,
-                func.coalesce(span_subq.c.span_seconds, 0) >= POOL_B_MIN_SPAN_DAYS * 86400,
-            ),
-        )
-        .order_by(DetectedEvent.confidence.desc())
-    )
-    for event, contact, span_seconds in b3_result.all():
-        if contact.id in queued_contact_ids:
-            continue
-        span_days = float(span_seconds or 0) / 86400.0
-        priority = compute_priority_b(
-            contact.interaction_count, contact.relationship_score, span_days, True,
-        )
-        if contact.id not in candidates or priority > candidates[contact.id].priority:
-            candidates[contact.id] = _Candidate(
-                contact=contact, trigger_type="dormant_event", event=event, priority=priority, pool="B",
-            )
-
-    # Apply hard cap: exclude contacts dormant > 5 years UNLESS they have a B3 event trigger
-    to_remove = []
-    for cid, cand in candidates.items():
-        if cand.trigger_type == "dormant_event":
-            continue  # B3 overrides hard cap
-        contact = cand.contact
-        if contact.last_interaction_at and contact.last_interaction_at < hard_cap_cutoff:
-            to_remove.append(cid)
-        elif contact.last_interaction_at is None:
-            to_remove.append(cid)
-    for cid in to_remove:
-        del candidates[cid]
-
-    return candidates
-
 
 # -----------------------------------------------------------------------
 # Main orchestrator
@@ -604,12 +290,17 @@ async def _generate_suggestions_inner(
 
     # Collect candidates from both pools. The user's configured dormancy
     # threshold defines the boundary: <= threshold = Pool A, > threshold = Pool B.
+    # Lazy import to avoid circular dep — followup_pools imports from us.
+    from app.services.followup_pools import (
+        _collect_pool_a_candidates,
+        _collect_pool_b_candidates,
+    )
+
     _prefs = settings.get("suggestion_prefs", {})
     dormancy_days = int(_prefs.get("dormancy_threshold_days", DORMANCY_THRESHOLD_DAYS))
     pool_a = await _collect_pool_a_candidates(
         user_id, db, now, queued_contact_ids, priority_settings, dormancy_days=dormancy_days,
     )
-    # Respect include_dormant preference — skip Pool B if disabled
     if _prefs.get("include_dormant", True):
         pool_b = await _collect_pool_b_candidates(
             user_id, db, now, queued_contact_ids, dormancy_days=dormancy_days,
@@ -792,22 +483,5 @@ async def _generate_suggestions_inner(
     return created
 
 
-async def get_weekly_digest(user_id: uuid.UUID, db: AsyncSession) -> list[FollowUpSuggestion]:
-    """Fetch pending suggestions for a user ordered by creation date (most recent first).
-
-    Args:
-        user_id: Target user's UUID.
-        db: Async database session.
-
-    Returns:
-        List of pending FollowUpSuggestion objects.
-    """
-    result = await db.execute(
-        select(FollowUpSuggestion)
-        .where(
-            FollowUpSuggestion.user_id == user_id,
-            FollowUpSuggestion.status == "pending",
-        )
-        .order_by(FollowUpSuggestion.created_at.desc())
-    )
-    return list(result.scalars().all())
+# Re-export for backwards compatibility — get_weekly_digest moved to followup_digest.py
+from app.services.followup_digest import get_weekly_digest  # noqa: E402,F401
